feat: remove config
